\name{agecalc}
\alias{agecalc}
\title{Calculates age}
\description{Tool to calculate ages from date of birth and date of test, etc}
\usage{agecalc( from, to )}
\arguments{
  \item{from}{date from}
  \item{to}{date to}
}

\value{returns age in years}

\examples{
  agecalc( "1977-01-31", "2014-01-30" )
}

\author{Ivan Marin-Franch <imarinfr@indiana.edu>}