\name{loadvfcsv}
\alias{loadvfcsv}
\title{loads visual fields from a CSV file}
\description{loads visual fields from a CSV file}
\usage{loadvfcsv( filename, patternMap )}
\arguments{
  \item{filename}{filename}
  \item{patternMap}{pattern of stimulus locations}
}
\details{
GIVE HERE HOW THE CSV FILE SHOULD BE FORMATTED
}

\value{returns a data frame in which each row is a visual field}

\references{
References?
}

\seealso{ \code{\link{loadvfxml}} }

\examples{
#loadvfcsv(  )
}

\author{Ivan Marin-Franch <imarinfr@indiana.edu>}