\name{ tdpmap }
\alias{ tdpmap }
\title{ probability map for total deviation }
\description{
gets probability map for total-deviation values
}
\usage{
tdpmap( td )
}
\arguments{
  \item{td}{ total-deviation values in dBs }
}
\details{
calculates the probability maps for total-deviation values
}

\value{
returns the subjects's total-deviation probability map from the age-matched normative reference.
}

\references{
References?
}

\seealso{ \code{\link{tdval}}, \code{\link{pdval}}, \code{\link{pdpmap}} }

\examples{
#tdpmap()
}

\author{Ivan Marin-Franch <imarinfr@indiana.edu>}