\name{tdrank}
\alias{tdrank}
\title{total-deviation rank curve}
\description{
get total-deviation rank curve
}
\usage{
tdrank( td )
}
\arguments{
  \item{td}{total deviation values}
}
\details{
calculates total-deviation rank curve.
}

\value{
returns the total-deviation rank curve
}

\references{
References?
}

\seealso{ \code{\link{tdrankperc}} }

\examples{
#tdrank(  )
}

\author{Ivan Marin-Franch <imarinfr@indiana.edu>}