\name{tdranknv}
\alias{tdranknv}
\title{normative values for TD rank curve}
\description{gets percentiles for TD rank curve}
\usage{tdranknv( td )}
\arguments{
  \item{td}{table with total-deviation values}
}
\details{details?}

\value{
percentiles for TD rank curves
}

\references{
References?
}

\seealso{ \code{\link{tdrank}} }

\examples{
#tdranknv()
}

\author{Ivan Marin-Franch <imarinfr@indiana.edu>}