\name{vfColorMap}
\alias{vfColorMap}
\title{Color Map for sensitivities}
\description{This function geneartes an rgb value to represents sensitivity at each location.}
\usage{vfColorMap( map )}
\arguments{
  \item{map}{deviation probability object}
}
\details{This function geneartes an rgb value to represents sensitivity at each location.}

\value{
something
}

\references{
something
}

\seealso{\code{\link{vfplot}}}

\examples{
#vfColorMap( map )
}

\author{Chaitanya Khadilkar <chaikhad@indiana.edu>, Ivan Marin-Franch <imarinfr@indiana.edu>}