\name{vfdemographics}
\alias{vfdemographics}
\title{demographics and statistics of sample in visual-fields object vf}
\description{gets demographics and weighted statistics of sample in visual-fields object vf}
\usage{ vfdemographics( vf ) }
\arguments{
  \item{vf}{A vf object. Tipically stimulus sensitivities, but can be total deviation values and pattern deviation values}
}

\value{returns a table with demographics and weighted statistics}

\references{
References?
}

\examples{
#vfdemographics(  )
}

\author{Ivan Marin-Franch <imarinfr@indiana.edu>}