\name{vfgrayscale}
\alias{vfgrayscale}
\title{Gray scales for sensitivities}
\description{This function maps sensitivity values to grayscales}
\usage{vfgrayscale( vf, age, pattern, algorithm )}
\arguments{
  \item{vf}{sensitivity values}
  \item{age}{age of the subject}
  \item{algorithm}{algorithm used in the visual test}
  \item{pattern}{pattern used in the visual test}
}
\details{This function maps sensitivity values to grayscales. The grayscale depends on age of the person, test pattern and algorithm
}

\value{
something
}

\references{
something
}

\seealso{\code{\link{vfColorMap}}}

\examples{
#vfgrayscale()
}

\author{Ivan Marin-Franch <imarinfr@indiana.edu>}