\name{vfindexpmap}
\alias{vfindexpmap}
\title{probability values for visual field index}
\description{calculates the probability values for visual field index}
\usage{ vfindexpmap( vfi ) }
\arguments{
  \item{vfi}{ visual fields index }
}
\details{
calculates the proability values for visual field index (see \code{\link{vfindex}})
}

\value{
returns the subjects's probability values for visual field index
}

\references{
References?
}

\seealso{ \code{\link{vfindex}} }

\examples{
#vfindexpmap(  )
}

\author{Ivan Marin-Franch <imarinfr@indiana.edu>}