\name{vfstatspmap}
\alias{vfstatspmap}
\title{probability values for visual field global indices}
\description{calculates the probability values for visual field global indices}
\usage{ vfstatspmap( vfindices ) }
\arguments{
  \item{vfindices}{ visual fields global indices }
}
\details{
calculates the proability values for visual field global indices (see \code{\link{vfstats}})
}

\value{
returns the subjects's probability values for visual field global indices
}

\references{
References?
}

\seealso{ \code{\link{vfstats}} }

\examples{
#vfstatspmap(  )
}

\author{Ivan Marin-Franch <imarinfr@indiana.edu>}