\name{visualFields-package}
\alias{visualFields}
\title{Statistical methods for visual fields}
\docType{package}
\description{
This is a collection of tools for analyzing the field of vision. It provides a framework for the development and use of innovative methods for visualization, statistical analysis, and clinical interpretation of visual-field loss and its change over time. \code{visualFields} is intended to be a tool for collaborative research
}
\details{This work was supported by NIH grant number \bold{R01EY007716}. We thank the following people who contributed to the package
  \tabular{ll}{
   \bold{Chaitanya Khadilkar} \tab developer\cr
   \bold{Neil O'Leary}        \tab provided the original implementation of PoPLR in MatLab\cr
   \bold{William H Swanson}   \tab provided data for control subjects for the generation of normative values\cr
   \bold{Victor E Malinovsky} \tab provided data for control subjects for the generation of normative values\cr
   \bold{Mitchell W Dul}      \tab provided data for control subjects for the generation of normative values\cr
   \bold{Richard Russell}     \tab provided some code to read XML files extracted from the HFA
   }
}
\references{ADD REFERENCES HERE}

\seealso{\code{OPI}}

\author{Ivan Marin-Franch <imarinfr@indiana.edu>, Paul H Artes <paul@dal.ca>}