\name{xmlitem}
\alias{xmlitem}
\title{extracts from a loaded XML file the info of a tag}
\description{extracts from a loaded XML file the info of a tag}
\usage{xmlitem( tag, xmllines )}
\arguments{
  \item{tag}{tag to look at}
  \item{xmllines}{lines from loaded XML files}
}
\details{
extracts from a loaded XML file the info of a tag
}

\value{returns the value of the tag}

\references{
References?
}

\seealso{ \code{\link{loadvfxml}}, \code{\link{loadvfcsv}}, \code{\link{xmlblock}} }

\examples{
#xmlitem(  )
}

\author{Ivan Marin-Franch <imarinfr@indiana.edu>}