\name{tdranknv}
\alias{tdranknv}
\title{normative values for TD rank curve}
\description{gets percentiles for TD rank curve}
\usage{tdranknv( td )}
\arguments{
  \item{td}{\code{\link{vf}}-object with total-deviation values}
}

\value{percentiles for TD rank curves}

\seealso{ \code{\link{tdrank}} }

\examples{
# DO NOT EXECUTE
#tdranknv( td )
}

\author{Ivan Marin-Franch <imarinfr@indiana.edu>}